fix(ci): resolve all clang-tidy violations in tests/mocks and add ci-rx72n target
Fix 537 unique clang-tidy violations across 54 test/mock files and 4 production
files that were caught by CI's ENABLE_CLANG_TIDY=ON CMake path (which checks all
compiled files including tests, unlike the local clang_tidy.sh script).

Add `make ci-rx72n` Makefile target to run the full local CI pipeline
(format, cross-compile, unit tests, clang-tidy) before pushing.

+++ star-rx72n-firmware/tests/mocks/mock_adc_hal.c
@@ -317,9 +317,8 @@ rx_err_t adc_init(adc_unit_t unit, adc_channel_t channel, adc_resolution_t bits)
   }
 
   /* Validate resolution */
-  if ((uint8_t)bits != (uint8_t)k_mock_adc_resolution_8bit &&
-      (uint8_t)bits != (uint8_t)k_mock_adc_resolution_10bit &&
-      (uint8_t)bits != (uint8_t)k_mock_adc_resolution_12bit) {
+  if ((uint8_t)bits != k_mock_adc_resolution_8bit && (uint8_t)bits != k_mock_adc_resolution_10bit &&
+      (uint8_t)bits != k_mock_adc_resolution_12bit) {
     return k_rx_err_invalid_arg;
   }
 
@@ -485,14 +484,14 @@ rx_err_t adc_read_voltage_mv(adc_unit_t       unit,
   }
 
   /* Validate resolution */
-  if ((uint8_t)bits != (uint8_t)k_mock_adc_resolution_8bit &&
-      (uint8_t)bits != (uint8_t)k_mock_adc_resolution_10bit &&
-      (uint8_t)bits != (uint8_t)k_mock_adc_resolution_12bit) {
+  if ((uint8_t)bits != k_mock_adc_resolution_8bit && (uint8_t)bits != k_mock_adc_resolution_10bit &&
+      (uint8_t)bits != k_mock_adc_resolution_12bit) {
     return k_rx_err_invalid_arg;
   }
 
   /* Validate resolution matches unit's configured resolution (matches production behavior) */
-  if ((uint8_t)unit < k_mock_adc_max_units && g_mock_adc.units[unit].initialized &&
+  if ((uint8_t)unit < k_mock_adc_max_units &&
+      g_mock_adc.units[unit].initialized && // NOLINT(readability-implicit-bool-conversion)
       g_mock_adc.units[unit].resolution != (uint8_t)bits) {
     return k_rx_err_invalid_arg;
   }
